requirements.txt
Pillow>=8.0.0
numpy>=1.19.0
click>=7.0
reportlab>=3.5.0